Write and serve JSON file
The json is served under /.well-known/nostr.json
It is written upon successful form submissions.

This doesn't scale well, because as part of the request-response cycle
the JSON file is rewritten from scratch each time. This slows down the
request. It is barely noticeable with a few entries in the DB.

Long term this could happen in a background job.