Convert keys between HEX and npub format
It is likely that users will post their keys in the 'npub...' format or
as a hex. Rather than work with errors and instruct the user, we can
solve this issue for them and just convert from the different formats
and then store them accordingly.

It's the hex version that will be published in the JSON.
We don't want to accept private keys ('nsec'). One way to deal with this
could be to convert any key starting with 'nsec' to the public key and
discard the private key.

To go from private to public we would need the ruby nostr library, which
currently only supports Ruby 3.3.