Write JSON file upon visit to homepage
My suspicion is that reploys remove the public file, so it has to be
created from scratch.

A temporary solution is to create it upon visiting the homepage and then
running regular health checks on it.

A more elegant alternative would be something like a cron job. However,
that incurs extra costs with the current dev ops setup.