Simplify serving of JSON
Now we just build and serve JSON from the controller. Previously we
would write the file first and then serve it from /public.

That had the advantage of it not being generated for each request,
however, there were numerous drawbacks:

The public file was overwritten on deploys with cleared cache and so no
nostr.json was available at all. As a workaround we created it upon
visiting the root_path, which was pinged regularly through healthchecks
from render.com. But that solution was brittle.