Handle name params
NIP-05 speaks of a name param.
https://github.com/nostr-protocol/nips/blob/master/05.md
The idea is that a request to e.g. /.well-known/nostr.json?name=jane
returns the hash with just the name of jane and the related public_key.

This commit adds handling of the param. Also, the case needs to be
handled when a username was requested that is not present.