Specify DATABASE_URL for test and development environments
When DATABASE_URL is set in the environment variables, Rails tries to
connect to that DB, even on test and development environments. That
fails, because the credentials are for the production database.

That's why this commit specifies DB URLs for test and development that
are used instead.

This way we can run tests and use the development database without
connection errors.

If it's necessary to load env variables, they can be added with:

$ export $(xargs <.env)