Set CORS header
NIP-05 states that the CORS header should be set to allow all origins.
Javascript apps might look up the npub of a name from a different
origin than the host of the app.

This commit uses rack-cors middleware to set the HTTP response header.
It's also added as a before-action.

Verify with:
$ curl -sI cosanostr.com/.well-known/nostr.json?name=robert | grep -i ^Access-Control
access-control-allow-origin: *